Pathfinding, Telemetry, and Vehicle Spawning Setup

+++ sim/build.sh
@@ -2,20 +2,20 @@
 
 # These commands build and install the sim project.
 
-safe_exit()
-{
-    if [[ $? == 1 ]]; then
-        exit 1
-    fi
-}
+set -e
 
 if [[ $1 == "-x" ]]; then
-    rm -rf build
+    
+    if [[ -d "./build" ]]; then
+        rm -rf build
+    fi
+
+    if [[ -d "./INSTALL" ]]; then
+        rm -rf INSTALL
+    fi
+
 fi
 
 cmake -S . -B build -DCMAKE_INSTALL_PREFIX="INSTALL"
-safe_exit
 cmake --build build --config Release -j$(nproc)
-safe_exit
 cmake --install build --config Release
-safe_exit
